Creates Network from SMILES
This program creates a network from SMILES based on the data format of

row1: smiles,name,chemblID,ProteinTarget (These are column labels)
the following rows are the data

The only data that needs to be populated within information are the smiles and ProteinTarget columns, the rest is discarded. The column names smiles and ProteinTarget need to be present. 

The output is a pairwise similarity network (smile1 smile2 simScore) that can be clustered using Markov clustering.

Additional improvements could be changing the data input requirements to only need the smiles format of the small molecule